minor fixes
code cleanup

+++ ardUI/src/Activity.cpp
@@ -141,8 +141,8 @@ void Activity::draw() const {
 
 void Activity::measure() {
 	if (rootView) {
-		int widthSpec = View::MeasureSpec::makeMeasureSpec(View::MeasureSpec::EXACTLY, arduiDisplayGetWidth());
-		int heightSpec = View::MeasureSpec::makeMeasureSpec(View::MeasureSpec::EXACTLY, arduiDisplayGetHeight());
+		uint16_t widthSpec = View::MeasureSpec::makeMeasureSpec(View::MeasureSpec::EXACTLY, arduiDisplayGetWidth());
+		uint16_t heightSpec = View::MeasureSpec::makeMeasureSpec(View::MeasureSpec::EXACTLY, arduiDisplayGetHeight());
 		rootView->measure(widthSpec, heightSpec);
 	}
 }
